Rohit/RAG_1/requirements.txt
# RAG_1 Project Dependencies

# Core RAG Framework
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-text-splitters==0.0.1

# Vector Store
faiss-cpu==1.7.4

# OpenAI API
openai==1.3.0

# Document Processing
python-docx==0.8.11
pypdf==3.17.1

# Utilities
python-dotenv==1.0.0
pydantic==2.0.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3

# Development
jupyter==1.0.0
ipython==8.19.0
